refactor(math): q_math predicates return bool — C truthiness call shape restored
VectorCompare/VectorCompare2/VectorBetweenVectors/PlaneFromPoints returned
qboolean/int per faithful transcription; Raven call sites use C implicit
truthiness (if (VectorCompare(..))), which the qboolean shape rendered as
!= 0 / == qfalse noise in Rust. Per porting-rules §C7 (user ruling, dedup
campaign) the four predicates now return bool and ~50 call sites across
bg/game/botlib/icarus read as the oracle does: if VectorCompare(..) / if
!VectorCompare(..). Int consumers cast at their site (icarus Q3_Evaluate
arm; jampgame_parity golden log keeps 0/1 via as i32). Inline
crate::q_math:: paths at touched sites extracted to prelude short names.

+++ crates/mp/game/tests/jampgame_parity.rs
@@ -349,7 +349,12 @@ fn dump_vecmath(o: &mut String, vecs: &[[f32; 3]]) {
             DistanceHorizontal(a, b).cbits(),
             DistanceHorizontalSquared(a, b).cbits()
         );
-        let _ = writeln!(o, "vcmp {} {}", VectorCompare(a, b), VectorCompare(a, a));
+        let _ = writeln!(
+            o,
+            "vcmp {} {}",
+            VectorCompare(a, b) as i32,
+            VectorCompare(a, a) as i32
+        );
         {
             let mut t = [0.0; 3];
             CrossProduct(a, b, &mut t);
@@ -441,7 +446,7 @@ fn dump_vecmath(o: &mut String, vecs: &[[f32; 3]]) {
         {
             let mut pl = [0.0; 4];
             let ok = PlaneFromPoints(&mut pl, a, b, c);
-            let _ = write!(o, "pfp {} ", ok);
+            let _ = write!(o, "pfp {} ", ok as i32);
             v4(o, &pl);
         }
         let _ = writeln!(o, "angsub {:08x}", AngleSubtract(a[0], a[1]).cbits());
